+++ vector/src/main/java/im/vector/riotx/features/home/room/detail/timeline/helper/TimelineDisplayableEvents.kt
@@ -27,6 +27,8 @@ object TimelineDisplayableEvents {
             EventType.STATE_ROOM_NAME,
             EventType.STATE_ROOM_TOPIC,
             EventType.STATE_ROOM_MEMBER,
+            EventType.STATE_ROOM_ALIASES,
+            EventType.STATE_CANONICAL_ALIAS,
             EventType.STATE_HISTORY_VISIBILITY,
             EventType.CALL_INVITE,
             EventType.CALL_HANGUP,